+++ pkg/razorpay/integrations/checkout_tool.go
@@ -1,0 +1,144 @@
+//nolint:lll // File contains embedded code templates requiring longer lines
+package integrations
+
+import (
+	"context"
+
+	rzpsdk "github.com/razorpay/razorpay-go"
+
+	"github.com/razorpay/razorpay-mcp-server/pkg/mcpgo"
+	"github.com/razorpay/razorpay-mcp-server/pkg/observability"
+)
+
+// IntegrateRazorpayCheckout returns a tool for complete Razorpay checkout integration
+//
+//nolint:gocyclo // Complex integration logic with multiple code paths
+func IntegrateRazorpayCheckout(
+	obs *observability.Observability,
+	client *rzpsdk.Client,
+) mcpgo.Tool {
+	parameters := []mcpgo.ToolParameter{
+		mcpgo.WithString(
+			"language",
+			mcpgo.Description("Programming language: javascript, typescript, python, go, java, php, ruby, rust, csharp, dart, swift, or kotlin"),
+			mcpgo.Required(),
+			mcpgo.Enum("javascript", "typescript", "python", "go", "java", "php", "ruby", "rust", "csharp", "dart", "swift", "kotlin"),
+		),
+		mcpgo.WithString(
+			"backendFramework",
+			mcpgo.Description("Backend framework: express, fastify, koa, nextjs, nuxt, django, flask, fastapi, gin, echo, fiber, spring, spring-boot, laravel, rails, actix, aspnet, react-native, flutter, android, ios, cordova, ionic, or capacitor"),
+			mcpgo.Required(),
+			mcpgo.Enum("express", "fastify", "koa", "nextjs", "nuxt", "django", "flask", "fastapi", "gin", "echo", "fiber", "spring", "spring-boot", "laravel", "rails", "actix", "aspnet", "react-native", "flutter", "android", "ios", "cordova", "ionic", "capacitor"),
+		),
+		mcpgo.WithString(
+			"frontendFramework",
+			mcpgo.Description("Frontend framework: vanilla, react, nextjs, vue, angular, svelte, or native (for mobile apps)"),
+			mcpgo.Required(),
+			mcpgo.Enum("vanilla", "react", "nextjs", "vue", "angular", "svelte", "native"),
+		),
+	}
+
+	handler := func(
+		ctx context.Context,
+		r mcpgo.CallToolRequest,
+	) (*mcpgo.ToolResult, error) {
+		args, ok := r.Arguments.(map[string]interface{})
+		if !ok {
+			return mcpgo.NewToolResultError("Invalid arguments"), nil
+		}
+
+		language, _ := args["language"].(string)
+		backendFramework, _ := args["backendFramework"].(string)
+		frontendFramework, _ := args["frontendFramework"].(string)
+
+		// Get client from context (for remote/SSE) or use default (for stdio)
+		activeClient, err := getClientFromContextOrDefault(ctx, client)
+		if err != nil {
+			return mcpgo.NewToolResultError("Failed to get client: " + err.Error()), nil
+		}
+
+		// Get credentials from client (passed via Authorization header or env vars)
+		creds := Credentials{
+			KeyID:     activeClient.Auth.Key,
+			KeySecret: activeClient.Auth.Secret,
+		}
+
+		var output IntegrateCheckoutOutput
+
+		// Get frontend code based on frontend framework
+		frontendCode := getFrontendIntegration(frontendFramework, creds)
+
+		// Route to appropriate backend integration
+		switch backendFramework {
+		// Python frameworks
+		case "django":
+			output = getDjangoIntegration(creds, frontendCode)
+		case "flask":
+			output = getFlaskIntegration(creds, frontendCode)
+		case "fastapi":
+			output = getFastAPIIntegration(creds, frontendCode)
+		// Go frameworks
+		case "gin":
+			output = getGinIntegration(creds, frontendCode)
+		case "echo":
+			output = getEchoIntegration(creds, frontendCode)
+		case "fiber":
+			output = getFiberIntegration(creds, frontendCode)
+		// Node.js frameworks
+		case "nextjs":
+			output = getNextjsReactIntegration(language, creds)
+		case "nuxt":
+			output = getNuxtIntegration(language, creds)
+		case "fastify":
+			output = getFastifyIntegration(language, creds, frontendCode)
+		case "koa":
+			output = getKoaIntegration(language, creds, frontendCode)
+		// Java
+		case "spring":
+			output = getSpringIntegration(creds, frontendCode)
+		case "spring-boot":
+			output = getSpringBootIntegration(creds, frontendCode)
+		// PHP
+		case "laravel":
+			output = getLaravelIntegration(creds, frontendCode)
+		// Ruby
+		case "rails":
+			output = getRailsIntegration(creds, frontendCode)
+		// Rust
+		case "actix":
+			output = getActixIntegration(creds, frontendCode)
+		// .NET
+		case "aspnet":
+			output = getAspNetIntegration(creds, frontendCode)
+		// Mobile
+		case "react-native":
+			output = getReactNativeIntegration(creds)
+		case "flutter":
+			output = getFlutterIntegration(creds)
+		case "android":
+			output = getAndroidIntegration(creds)
+		case "ios":
+			output = getIOSIntegration(creds)
+		case "cordova":
+			output = getCordovaIntegration(creds)
+		case "ionic":
+			output = getIonicIntegration(creds)
+		case "capacitor":
+			output = getCapacitorIntegration(creds)
+		default: // express
+			output = getExpressVanillaIntegration(language, creds, frontendCode)
+		}
+
+		return mcpgo.NewToolResultJSON(output)
+	}
+
+	return mcpgo.NewTool(
+		"integrate_razorpay_checkout",
+		"Complete Razorpay Standard Checkout integration. Returns ALL code needed - "+
+			"backend routes, frontend integration, and payment verification. "+
+			"Use this single tool to get everything needed for Razorpay payment integration. "+
+			"The AI should apply ALL returned files and modifications without asking the user for additional steps.",
+		parameters,
+		handler,
+	)
+}
